+++ docs/transactions.rst
@@ -1,10 +1,13 @@
 Transactions
 ============
 
-.. py:module:: starknet_py.transactions.deploy
+.. warning::
+    This module is deprecated and will be removed in the future versions.
+
+.. py:module:: starknet_py.net.transactions.deploy
 
 .. autofunction:: make_deploy_tx
 
-.. py:module:: starknet_py.transactions.declare
+.. py:module:: starknet_py.net.transactions.declare
 
 .. autofunction:: make_declare_tx
