+++ starknet_py/tests/e2e/client/client_test.py
@@ -26,8 +26,8 @@ from starknet_py.transaction_exceptions import (
     TransactionRejectedError,
     TransactionNotReceivedError,
 )
-from starknet_py.transactions.declare import make_declare_tx
-from starknet_py.transactions.deploy import make_deploy_tx
+from starknet_py.net.transactions.declare import make_declare_tx
+from starknet_py.net.transactions.deploy import make_deploy_tx
 
 
 @pytest.mark.asyncio
